feat: use OTEL for tracing instead of Sentry, add riverqueue OTEL middleware (#109)
This PR initializes an OTEL Tracer and export spans to Sentry, so traces will also be sent to Sentry in addition to an OTEL endpoint.
Also, existing Sentry spans are not created any more. Instead, root
spans are created by the riverqueue OTEL middleware (riverqueue.com/docs/open-telemetry). And a child span is created for the call to each module in `internal/background/modules_worker/worker.go` when processing a new Slack message.

When reporting an error to Sentry, the `Context` is added as a hint to
attach the trace context to the error.

Also, add instructions to README.md on viewing traces in dev environment
in either sentry.io or a local Grafana Tempo instance.