feat: trace agent message generate command with force sampler (#111)
* feat: trace agent message generate command with force sampler

Add a custom OTEL span sampler that will always sample a span with the
`force_trace` attribute and set that attribute on the `Generate` and
`Respond` commands if `force` is set to true.

Also add the OTEL HTTP handler to the command routes for testing in dev.

And start centralizing attribute names in the `semconv` package.